initial backport of native double tap to wake from Xperia Z2

+++ include/linux/clearpad.h
@@ -18,7 +18,6 @@
 #include <linux/kernel.h>
 #include <linux/of.h>
 #include <linux/platform_device.h>
-#include <linux/input/evgen_helper.h>
 
 #define CLEARPAD_NAME "clearpad"
 #define CLEARPADI2C_NAME "clearpad-i2c"
@@ -66,7 +65,7 @@ struct synaptics_button_data {
 struct synaptics_easy_wakeup_config {
 	bool gesture_enable;
 	bool large_panel;
-	unsigned long timeout_delay;
+	u32 timeout_delay;
 };
 
 struct clearpad_platform_data {
